+++ client/src/pages/DevelopmentMode.tsx
@@ -1,6 +1,5 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { socket } from 'utils/socket';
-import CodeEditor from 'components/DevelopmentMode/CodeEditor';
 import styles from 'styles/DevelopmentMode.module.css';
 import globalStyles from 'styles/global.module.css';
 import classNames from 'classnames';
